Moving new classes under index.codec.util
