+++ src/test/java/org/opensearch/knn/TestUtils.java
@@ -15,7 +15,7 @@ import org.opensearch.common.xcontent.DeprecationHandler;
 import org.opensearch.common.xcontent.NamedXContentRegistry;
 import org.opensearch.common.xcontent.XContentFactory;
 import org.opensearch.common.xcontent.XContentType;
-import org.opensearch.knn.index.codec.KNNCodecUtil;
+import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
